+++ src/main/kotlin/com/hyungjun/todolist/task/entity/TaskMemoEntity.kt
@@ -8,7 +8,7 @@ import java.time.LocalDateTime
 @Table(name = "TaskMemos")
 data class TaskMemoEntity(
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-	val id: Int = 0,
+	val id: Long = 0,
 
 	@ManyToOne
 	@JoinColumn(name = "user_id", nullable = false)
